feat: 添加模型分组支持，更新相关DTO和前端表单，增强流程模型管理功能

+++ src/main/java/com/ksptool/bio/biz/qf/controller/QfModelController.java
@@ -69,6 +69,12 @@ public class QfModelController {
     @Operation(summary = "删除流程模型")
     @PostMapping("/removeQfModel")
     public Result<String> removeQfModel(@RequestBody @Valid CommonIdDto dto) throws Exception {
+
+        //不支持批量删除
+        if (dto.isBatch()) {
+            return Result.error("不支持批量删除");
+        }
+
         qfModelService.removeQfModel(dto);
         return Result.success("操作成功");
     }
